Refactor transaction.py final phase
