+++ counterparty-core/counterpartycore/lib/transaction_helper/common_serializer.py
